repository class

+++ src/main/java/com/sample/gallery/repository/ImageGalleryRepository.java
@@ -2,6 +2,8 @@ package com.sample.gallery.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageGalleryRepository extends JpaRepository<ImageGallery,Integer>{
+import com.sample.gallery.entity.ImageGallery;
+
+public interface ImageGalleryRepository extends JpaRepository<ImageGallery, Long>{
     
 }
